Fix compressiong old period logs (previously not compressed)
Old period logs (e.g. name.201311, name.201312, etc) do not have suffix and thus
were never compressed. We now check the period against current period. The
solution is a bit hackish currently: file_path() is made to return the current
period as the second element of list, and _open() saves it in the _cur_period
attribute.